Design changes.
New design for season view, it's now a picker with a horizontal list of episodes below it.
New style for list view used in Explore, Watchlist, and Search.
HomeView now has to new horizontal list, one shows upcoming movies that have a date and notification is set to true, and the other list is for upcoming new tv shows' seasons.