Developer View
DeveloperView is a view that helps test some features, it can fetch ItemContent and Person objects and get a list of pending notifications.
The main purpose of this view is to develop some features and test specific objects.
Added fetchUpcomingNotifications() to NotificationManager, this method will return every notification as ItemContent objects.