Improved VoiceOver support.
Added function to check if an item is archived.
Changed BackgroundManager to use newBackgroundContext.
Added basic validation to check if an item has been recently updated, if so it will skip the item in the sync.
Added a new view, DetailedPeopleList, to see a vertical list of every cast and crew.
Limited the CastListView to display only 10 items.