Fixed contentID.
This commit marks the beginning of the rewrite for some essential functions in the app, including those related to Core Data. The purpose of this rewrite is to establish a stronger foundation for adding new features in a more sustainable manner.

In this commit, I've made significant changes to the fetch request used for retrieving WatchlistItems. Previously, the function required both the item's ID and media type, and then it would fetch items with the same ID and compare the media types to determine if a match was found.

Now, instead of separate ID and media type inputs, the function only requires a unified string, called itemNotificationID. This string represents an item's ID followed by its media type as an integer, like this: 455476@0. The use of this unified string eliminates the need for comparing after fetching, resulting in a more streamlined process.