2.1.3
Added favorite, pin, and archive buttons to watchOS.
ChangelogView now has a frame for macOS when it opens the Tip Jar view.
Added styleButton for custom lists.
Removed notification status for ItemContentDetails.
NotificationListView will now check if an item is removed from the watchlist, if so, it'll remove the notification.
ChangelogView is now deployed, frame modifier has been added to the macOS version.
Removed more actions menu for macOS on the season list view.
Theme picker now uses .segmented style.
AppIcon Selector is now a var inside the AppearanceSettings view and displays in-line with the rest of the view.
Removed FeatureRoadmap until API for counting votes is back-up again.