Improved support for iOS and iPadOS 18.
Fixed the toolbar items not activating.
Improved the usage of material by forcing the color scheme to dark to enhance legibility.
Moved the corner radius to 12 to make it easier to maintain the codebase with tvOS.
Improved the wording when the recommendation is empty on the Discover tab.
Added the "Huge Tip" option to the UI.
Changed the behavior of the Tip screen to allow users to tip more if they want to.
Added option to users clean the cache.
Almost all changes here only apply to iOS and iPadOS; I haven't tested them yet on visionOS and watchOS.